Decompose spec.c into spec + indexes + indexes_scan
src/spec.c had grown to ~5k lines tangling three independently
reasoned-about subsystems. Split it into three translation units with a
strictly one-directional dependency graph (no cyclic deps, no umbrella
header):

- spec.{c,h}        IndexSpec lifecycle: create/alter/free, field/GC/
                    cursor wiring, single-spec RDB.
- indexes.{c,h}     Global registry + keyspace dispatch over specDict_g/
                    specIdDict_g: RDB load/save, propagation, the
                    Indexes_*MatchingWithSchemaRules family, RDB-load
                    lifecycle events, and SST replication.
- indexes_scan.{c,h} Background scanner + reindex: IndexesScanner family,
                    the debug scanner, reindexPool, full-rescan and
                    legacy-index upgrade entry points.

Header hygiene: spec.h depends on neither new header; indexes_scan.h
forward-declares IndexSpec (never includes spec.h); indexes.h includes
spec.h but not indexes_scan.h. The indexes.c -> indexes_scan.h and
*.c -> spec.h edges live only in the .c files.

Break the one function-call cycle (spec -> scanner) by hoisting the
initial-scan trigger to the caller. A new Indexes_CreateNew wrapper in
the registry layer composes IndexSpec_CreateNew + IndexSpec_ScanAndReindex;
module.c and the cpptest call sites use it, so behavior is unchanged.
IndexSpec_AddFields drops its initialScan parameter and the scan is
scheduled by module.c's ALTER path instead. spec.c retains only a
data-only read of global_spec_scanner in IndexSpec_AddToInfo.

specDict_g/specIdDict_g definitions move to indexes.c; their extern
declarations stay in spec.h since the IndexSpec lifecycle reads them as a
data dependency. CheckRdbSstPersistence, IndexSpec_CreateFromRdb,
IndexSpec_Set/ResetTimeoutTimer and IndexSpec_SSTRdbOpenAndApply are
exposed (de-static'd) so the registry can call them.

FFI: move the Indexes_Init bindgen allowlist entry from spec.h to the new
indexes.h in ffi/build.rs; specDict_g/specIdDict_g stay under spec.h.

Pure code movement plus one behavior-preserving hoist: no new behavior,
no RDB format change, no change to scanner or keyspace-dispatch logic.
