+++ src/spec.c
@@ -7,6 +7,8 @@
  * GNU Affero General Public License v3 (AGPLv3).
 */
 #include "spec.h"
+#include "indexes.h"
+#include "indexes_scan.h"
 #include "document.h"
 #include "inverted_index_ffi.h"
 #include "numeric_range_tree_ffi.h"
@@ -56,23 +58,20 @@
 #include "search_disk_utils.h"
 #include "iterators_ffi.h"
 
-#define INITIAL_DOC_TABLE_SIZE 1000
-
 ///////////////////////////////////////////////////////////////////////////////////////////////
 
 const char *(*IndexAlias_GetUserTableName)(RedisModuleCtx *, const char *) = NULL;
 
 RedisModuleType *IndexSpecType;
 
-dict *specDict_g = NULL;
-dict *specIdDict_g = NULL;
-IndexesScanner *global_spec_scanner = NULL;
+// The global index registry (specDict_g / specIdDict_g) is defined in indexes.c.
+// Its extern declarations live in spec.h, since the IndexSpec lifecycle functions
+// here read it as a data dependency.
 
 // Maximum number of indexes that can be created.
 // Can be modified via FT.DEBUG for testing.
 uint32_t maxIndexes_g = DEFAULT_MAX_INDEXES;
 
-size_t pending_global_indexing_ops = 0;
 dict *legacySpecDict;
 dict *legacySpecRules;
 
@@ -99,22 +98,6 @@ static redisearch_thpool_t *cleanPool = NULL;
 
 extern DebugCTX globalDebugCtx;
 
-const char *DEBUG_INDEX_SCANNER_STATUS_STRS[] = {
-    "NEW", "SCANNING", "DONE", "CANCELLED", "PAUSED", "RESUMED", "PAUSED_ON_OOM", "PAUSED_BEFORE_OOM_RETRY",
-};
-
-// Static assertion to ensure array size matches the number of statuses
-static_assert(
-    (sizeof(DEBUG_INDEX_SCANNER_STATUS_STRS) / sizeof(char*)) == DEBUG_INDEX_SCANNER_CODE_COUNT,
-    "Mismatch between DebugIndexScannerCode enum and DEBUG_INDEX_SCANNER_STATUS_STRS array"
-);
-
-// Debug scanner functions
-static DebugIndexesScanner *DebugIndexesScanner_New(StrongRef global_ref);
-static void DebugIndexesScanner_Free(DebugIndexesScanner *dScanner);
-static void DebugIndexes_ScanProc(RedisModuleCtx *ctx, RedisModuleString *keyname, RedisModuleKey *key,
-                             DebugIndexesScanner *dScanner);
-static void DebugIndexesScanner_pauseCheck(DebugIndexesScanner* dScanner, RedisModuleCtx *ctx, bool pauseField, DebugIndexScannerCode code);
 
 // Forward declaration for disk validation
 inline static bool isSpecOnDiskForValidation(const IndexSpec *sp);
@@ -130,46 +113,6 @@ bool CheckRdbSstPersistence(RedisModuleCtx *ctx, const char* prefix) {
 
 //---------------------------------------------------------------------------------------------
 
-// This function should be called after the first background scan OOM error
-// It will wait for resource manager to allocate more memory to the process if possible
-// and after the function returns, the scan will continue
-static inline void threadSleepByConfigTime(RedisModuleCtx *ctx, IndexesScanner *scanner) {
-  // Thread sleep based on the config
-  uint32_t sleepTime = RSGlobalConfig.bgIndexingOomPauseTimeBeforeRetry;
-  RedisModule_Log(ctx, "notice", "Scanning index %s in background: paused for %u seconds due to OOM, waiting for memory allocation",
-                  scanner->spec_name_for_logs, sleepTime);
-
-  RedisModule_ThreadSafeContextUnlock(ctx);
-  sleep(sleepTime);
-  RedisModule_ThreadSafeContextLock(ctx);
-}
-
-// This function should be called after the second background scan OOM error
-// It will stop the background scan process
-static inline void scanStopAfterOOM(RedisModuleCtx *ctx, IndexesScanner *scanner) {
-  char* error;
-  rm_asprintf(&error, "Used memory is more than %u percent of max memory, cancelling the scan", RSGlobalConfig.indexingMemoryLimit);
-  RedisModule_Log(ctx, "warning", "%s", error);
-
-    // We need to report the error message besides the log, so we can show it in FT.INFO
-  if(!scanner->global) {
-    scanner->cancelled = true;
-    StrongRef curr_run_ref = WeakRef_Promote(scanner->spec_ref);
-    IndexSpec *sp = StrongRef_Get(curr_run_ref);
-    if (sp) {
-      sp->scan_failed_OOM = true;
-      // Error message does not contain user data
-      IndexError_AddError(&sp->stats.indexError, error, error, scanner->OOMkey);
-      IndexError_RaiseBackgroundIndexFailureFlag(&sp->stats.indexError);
-      StrongRef_Release(curr_run_ref);
-    } else {
-      // spec was deleted
-      RedisModule_Log(ctx, "notice", "Scanning index %s in background: cancelled due to OOM and index was dropped",
-                    scanner->spec_name_for_logs);
-      }
-    }
-    rm_free(error);
-}
 
 
 static void setMemoryInfo(RedisModuleCtx *ctx) {
@@ -188,18 +131,6 @@ static void setMemoryInfo(RedisModuleCtx *ctx) {
   RedisModule_FreeServerInfo(ctx, info);
 }
 
-// Return true if used_memory exceeds (indexingMemoryLimit % × memoryLimit); false if within bounds or limit is 0.
-static inline bool isBgIndexingMemoryOverLimit(RedisModuleCtx *ctx) {
-  // if memory limit is set to 0, we don't need to check for memory usage
-  if(RSGlobalConfig.indexingMemoryLimit == 0) {
-    return false;
-  }
-
-  float used_memory_ratio = RedisMemory_GetUsedMemoryRatioUnified(ctx);
-  float memory_limit_ratio = (float)RSGlobalConfig.indexingMemoryLimit / 100;
-
-  return (used_memory_ratio > memory_limit_ratio) ;
-}
 /*
  * Initialize the spec's fields that are related to the cursors.
  */
@@ -411,7 +342,7 @@ static void IndexSpec_TimedOutProc(RedisModuleCtx *ctx, WeakRef w_ref) {
 
 // Assuming the GIL is held.
 // This can be done without locking the spec for write, since the timer is not modified or read by any other thread.
-static void IndexSpec_SetTimeoutTimer(IndexSpec *sp, WeakRef spec_ref) {
+void IndexSpec_SetTimeoutTimer(IndexSpec *sp, WeakRef spec_ref) {
   if (sp->isTimerSet) {
     WeakRef old_timer_ref;
     if (RedisModule_StopTimer(RSDummyContext, sp->timerId, (void **)&old_timer_ref) == REDISMODULE_OK) {
@@ -424,7 +355,7 @@ static void IndexSpec_SetTimeoutTimer(IndexSpec *sp, WeakRef spec_ref) {
 }
 
 // Assuming the spec is properly guarded before calling this function (GIL or write lock).
-static void IndexSpec_ResetTimeoutTimer(IndexSpec *sp) {
+void IndexSpec_ResetTimeoutTimer(IndexSpec *sp) {
   if (sp->isTimerSet) {
     WeakRef old_timer_ref;
     if (RedisModule_StopTimer(RSDummyContext, sp->timerId, (void **)&old_timer_ref) == REDISMODULE_OK) {
@@ -435,37 +366,9 @@ static void IndexSpec_ResetTimeoutTimer(IndexSpec *sp) {
   sp->isTimerSet = false;
 }
 
-// Assuming the GIL is locked before calling this function.
-void Indexes_SetTempSpecsTimers(TimerOp op) {
-  dictIterator *iter = dictGetIterator(specDict_g);
-  dictEntry *entry = NULL;
-  while ((entry = dictNext(iter))) {
-    StrongRef spec_ref = dictGetRef(entry);
-    IndexSpec *sp = StrongRef_Get(spec_ref);
-    if (sp->flags & Index_Temporary) {
-      switch (op) {
-        case TimerOp_Add: IndexSpec_SetTimeoutTimer(sp, StrongRef_Demote(spec_ref)); break;
-        case TimerOp_Del: IndexSpec_ResetTimeoutTimer(sp);    break;
-      }
-    }
-  }
-  dictReleaseIterator(iter);
-}
 
 //---------------------------------------------------------------------------------------------
 
-double IndexesScanner_IndexedPercent(RedisModuleCtx *ctx, IndexesScanner *scanner, const IndexSpec *sp) {
-  if (scanner || sp->scan_in_progress) {
-    if (scanner) {
-      size_t totalKeys = RedisModule_DbSize(ctx);
-      return totalKeys > 0 ? (double)scanner->scannedKeys / totalKeys : 0;
-    } else {
-      return 0;
-    }
-  } else {
-    return 1.0;
-  }
-}
 
 size_t IndexSpec_collect_numeric_overhead(IndexSpec *sp) {
   // Traverse the fields and calculates the overhead of the numeric tree index
@@ -610,9 +513,9 @@ IndexSpec *IndexSpec_CreateNew(RedisModuleCtx *ctx, RedisModuleString **argv, in
   // spec
   Initialize_KeyspaceNotifications();
 
-  if (!(sp->flags & Index_SkipInitialScan)) {
-    IndexSpec_ScanAndReindex(ctx, spec_ref);
-  }
+  // NOTE: kicking off the initial background scan is the caller's responsibility
+  // (see Indexes_CreateNew in indexes.c). Keeping it out of the IndexSpec core
+  // removes the spec -> scanner dependency.
   return sp;
 }
 
@@ -1693,17 +1596,14 @@ reset:
   return 0;
 }
 
-// Assumes the spec is locked for write
-int IndexSpec_AddFields(StrongRef spec_ref, IndexSpec *sp, RedisModuleCtx *ctx, ArgsCursor *ac, bool initialScan,
+// Assumes the spec is locked for write.
+// NOTE: scheduling the post-alter background scan is the caller's responsibility
+// (see CreateIndexAlterCommand in module.c), keeping spec.c free of a scanner edge.
+int IndexSpec_AddFields(StrongRef spec_ref, IndexSpec *sp, RedisModuleCtx *ctx, ArgsCursor *ac,
                         QueryError *status) {
   setMemoryInfo(ctx);
 
-  int rc = IndexSpec_AddFieldsInternal(sp, spec_ref, ac, status, 0);
-  if (rc && initialScan) {
-    IndexSpec_ScanAndReindex(ctx, spec_ref);
-  }
-
-  return rc;
+  return IndexSpec_AddFieldsInternal(sp, spec_ref, ac, status, 0);
 }
 
 bool IndexSpec_IsCoherent(IndexSpec *spec, sds* prefixes, size_t n_prefixes) {
@@ -2206,39 +2106,6 @@ void IndexSpec_RemoveFromGlobals(StrongRef spec_ref, bool removeActive) {
   StrongRef_Release(spec_ref);
 }
 
-void Indexes_Free(RedisModuleCtx *ctx, dict *d, bool deleteDiskData) {
-  // free the schema dictionary this way avoid iterating over it for each combination of
-  // spec<-->prefix
-  SchemaPrefixes_Free(SchemaPrefixes_g);
-  SchemaPrefixes_g = NULL;
-  SchemaPrefixes_Create();
-
-  CursorList_Empty(&g_CursorsListCoord);
-  // cursor list is iterating through the list as well and consuming a lot of CPU
-  CursorList_Empty(&g_CursorsList);
-
-  arrayof(StrongRef) specs = array_new(StrongRef, dictSize(d));
-  dictIterator *iter = dictGetIterator(d);
-  dictEntry *entry = NULL;
-  while ((entry = dictNext(iter))) {
-    StrongRef spec_ref = dictGetRef(entry);
-    array_append(specs, spec_ref);
-  }
-  dictReleaseIterator(iter);
-
-  for (size_t i = 0; i < array_len(specs); ++i) {
-    IndexSpec *spec = StrongRef_Get(specs[i]);
-    if (spec && spec->diskSpec) {
-      // Unregister must always precede close (triggered by IndexSpec_RemoveFromGlobals)
-      SearchDisk_UnregisterIndex(ctx, spec);
-      if (deleteDiskData) {
-        SearchDisk_MarkIndexForDeletion(spec->diskSpec);
-      }
-    }
-    IndexSpec_RemoveFromGlobals(specs[i], false);
-  }
-  array_free(specs);
-}
 
 
 //---------------------------------------- atomic updates ---------------------------------------
@@ -2818,293 +2685,6 @@ static void IndexStats_RdbLoad(RedisModuleIO *rdb, IndexStats *stats, int encver
 
 ///////////////////////////////////////////////////////////////////////////////////////////////
 
-static redisearch_thpool_t *reindexPool = NULL;
-
-static IndexesScanner *IndexesScanner_NewGlobal() {
-  if (global_spec_scanner) {
-    return NULL;
-  }
-
-  IndexesScanner *scanner = rm_calloc(1, sizeof(IndexesScanner));
-  scanner->global = true;
-  scanner->scannedKeys = 0;
-
-  global_spec_scanner = scanner;
-  RedisModule_Log(RSDummyContext, "notice", "Global scanner created");
-
-  return scanner;
-}
-
-static IndexesScanner *IndexesScanner_New(StrongRef global_ref) {
-
-  IndexesScanner *scanner = rm_calloc(1, sizeof(IndexesScanner));
-
-  scanner->spec_ref = StrongRef_Demote(global_ref);
-  IndexSpec *spec = StrongRef_Get(global_ref);
-  scanner->spec_name_for_logs = rm_strdup(IndexSpec_FormatName(spec, RSGlobalConfig.hideUserDataFromLog));
-  scanner->isDebug = false;
-
-  // scan already in progress?
-  if (spec->scanner) {
-    // cancel ongoing scan, keep on_progress indicator on
-    IndexesScanner_Cancel(spec->scanner);
-    const char* name = IndexSpec_FormatName(spec, RSGlobalConfig.hideUserDataFromLog);
-    RedisModule_Log(RSDummyContext, "notice", "Scanning index %s in background: cancelled and restarted", name);
-  }
-  spec->scanner = scanner;
-  spec->scan_in_progress = true;
-
-  return scanner;
-}
-
-void IndexesScanner_Free(IndexesScanner *scanner) {
-  rm_free(scanner->spec_name_for_logs);
-  if (global_spec_scanner == scanner) {
-    global_spec_scanner = NULL;
-  } else {
-    StrongRef tmp = WeakRef_Promote(scanner->spec_ref);
-    IndexSpec *spec = StrongRef_Get(tmp);
-    if (spec) {
-      if (spec->scanner == scanner) {
-        spec->scanner = NULL;
-        spec->scan_in_progress = false;
-      }
-      StrongRef_Release(tmp);
-    }
-    WeakRef_Release(scanner->spec_ref);
-  }
-  // Free the last scanned key
-  if (scanner->OOMkey) {
-    RedisModule_FreeString(RSDummyContext, scanner->OOMkey);
-  }
-  rm_free(scanner);
-}
-
-void IndexesScanner_Cancel(IndexesScanner *scanner) {
-  scanner->cancelled = true;
-}
-
-void IndexesScanner_ResetProgression(IndexesScanner *scanner) {
-  scanner-> scanFailedOnOOM = false;
-  scanner-> scannedKeys = 0;
-}
-
-//---------------------------------------------------------------------------------------------
-
-static void IndexSpec_DoneIndexingCallabck(struct RSAddDocumentCtx *docCtx, RedisModuleCtx *ctx,
-                                           void *pd) {
-}
-
-//---------------------------------------------------------------------------------------------
-
-int IndexSpec_UpdateDoc(IndexSpec *spec, RedisModuleCtx *ctx, RedisModuleString *key, DocumentType type);
-static void Indexes_ScanProc(RedisModuleCtx *ctx, RedisModuleString *keyname, RedisModuleKey *key,
-                             IndexesScanner *scanner) {
-
-  if (scanner->cancelled) {
-    return;
-  }
-
-  if (isBgIndexingMemoryOverLimit(ctx)){
-    scanner->scanFailedOnOOM = true;
-    if (scanner->OOMkey) {
-      RedisModule_FreeString(RSDummyContext, scanner->OOMkey);
-    }
-    // Hold the key that triggered OOM in case we need to attach an index error
-    scanner->OOMkey = RedisModule_HoldString(RSDummyContext, keyname);
-    return;
-  }
-  // RMKey it is provided as best effort but in some cases it might be NULL
-  bool keyOpened = false;
-  if (!key || isCrdt) {
-    key = RedisModule_OpenKey(ctx, keyname, DOCUMENT_OPEN_KEY_INDEXING_FLAGS);
-    keyOpened = true;
-  }
-  // Get the document type
-  DocumentType type = getDocType(key);
-
-  // Close the key if we opened it
-  if (keyOpened) {
-    RedisModule_CloseKey(key);
-  }
-
-  // Verify that the document type is supported and document is not empty
-  if (type == DocumentType_Unsupported) {
-    return;
-  }
-
-  if (scanner->global) {
-    Indexes_UpdateMatchingWithSchemaRules(ctx, keyname, type, NULL);
-  } else {
-    StrongRef curr_run_ref = IndexSpecRef_Promote(scanner->spec_ref);
-    IndexSpec *sp = StrongRef_Get(curr_run_ref);
-    if (sp) {
-      // This check is performed without locking the spec, but it's ok since we locked the GIL
-      // So the main thread is not running and the GC is not touching the relevant data
-      if (SchemaRule_ShouldIndex(sp, keyname, type)) {
-        IndexSpec_UpdateDoc(sp, ctx, keyname, type);
-      }
-      IndexSpecRef_Release(curr_run_ref);
-    } else {
-      // spec was deleted, cancel scan
-      scanner->cancelled = true;
-    }
-  }
-  ++scanner->scannedKeys;
-}
-
-//---------------------------------------------------------------------------------------------
-// Define for neater code, first argument is the debug scanner flag field , second is the status code
-#define IF_DEBUG_PAUSE_CHECK(scanner, ctx, status_bool, status_code) \
-if (scanner->isDebug) { \
-  DebugIndexesScanner *dScanner = (DebugIndexesScanner*)scanner;\
-  DebugIndexesScanner_pauseCheck(dScanner, ctx, dScanner->status_bool, status_code); \
-}
-#define IF_DEBUG_PAUSE_CHECK_BEFORE_OOM_RETRY(scanner, ctx) IF_DEBUG_PAUSE_CHECK(scanner, ctx, pauseBeforeOOMRetry, DEBUG_INDEX_SCANNER_CODE_PAUSED_BEFORE_OOM_RETRY)
-#define IF_DEBUG_PAUSE_CHECK_ON_OOM(scanner, ctx) IF_DEBUG_PAUSE_CHECK(scanner, ctx, pauseOnOOM, DEBUG_INDEX_SCANNER_CODE_PAUSED_ON_OOM)
-
-static void Indexes_ScanAndReindexTask(IndexesScanner *scanner) {
-  RS_LOG_ASSERT(scanner, "invalid IndexesScanner");
-
-  size_t counter = 0;
-  RedisModuleScanCB scanner_func = (RedisModuleScanCB)Indexes_ScanProc;
-
-  RedisModuleCtx *ctx = RedisModule_GetDetachedThreadSafeContext(RSDummyContext);
-  RedisModuleScanCursor *cursor = RedisModule_ScanCursorCreate();
-  RedisModule_ThreadSafeContextLock(ctx);
-
-  if (scanner->cancelled) {
-    goto end;
-  }
-  if (scanner->global) {
-    RedisModule_Log(ctx, "notice", "Scanning indexes in background");
-  } else {
-    RedisModule_Log(ctx, "notice", "Scanning index %s in background", scanner->spec_name_for_logs);
-  }
-  if (globalDebugCtx.debugMode) {
-    // If we are in debug mode, we need to use the debug scanner function
-    scanner_func = (RedisModuleScanCB)DebugIndexes_ScanProc;
-
-    // If background indexing paused, wait until it is resumed
-    // Allow the redis server to acquire the GIL while we release it
-    RedisModule_ThreadSafeContextUnlock(ctx);
-    while (globalDebugCtx.bgIndexing.pause) { // volatile variable
-      usleep(1000);
-    }
-    RedisModule_ThreadSafeContextLock(ctx);
-  }
-
-  while (RedisModule_Scan(ctx, cursor, scanner_func, scanner)) {
-    RedisModule_ThreadSafeContextUnlock(ctx);
-    counter++;
-    if (counter % RSGlobalConfig.numBGIndexingIterationsBeforeSleep == 0) {
-      // Sleep to allow redis server to acquire the GIL while we release it.
-      // We do that periodically every X iterations (100 as default), otherwise we call
-      // 'sched_yield()'. That is since 'sched_yield()' doesn't give up the processor for enough
-      // time to ensure that other threads that are waiting for the GIL will actually have the
-      // chance to take it.
-      usleep(RSGlobalConfig.bgIndexingSleepDurationMicroseconds);
-      IncrementBgIndexYieldCounter();
-    } else {
-      sched_yield();
-    }
-    RedisModule_ThreadSafeContextLock(ctx);
-
-    // Check if we need to handle OOM but must check if the scanner was cancelled for other reasons (i.e. FT. ALTER)
-    if (scanner->scanFailedOnOOM && !scanner->cancelled) {
-
-      // Check the config to see if we should wait for memory allocation
-      if(RSGlobalConfig.bgIndexingOomPauseTimeBeforeRetry > 0) {
-        IF_DEBUG_PAUSE_CHECK_BEFORE_OOM_RETRY(scanner, ctx);
-        // Call the wait function
-        threadSleepByConfigTime(ctx, scanner);
-        if (!isBgIndexingMemoryOverLimit(ctx)) {
-          // We can continue the scan
-          RedisModule_Log(ctx, "notice", "Scanning index %s in background: resuming after OOM due to memory limit increase",
-                          scanner->spec_name_for_logs);
-          IndexesScanner_ResetProgression(scanner);
-          RedisModule_ScanCursorRestart(cursor);
-          continue;
-        }
-      }
-      // At this point we either waited for memory allocation and failed
-      // or the config is set to not wait for memory allocation after OOM
-      scanStopAfterOOM(ctx, scanner);
-      IF_DEBUG_PAUSE_CHECK_ON_OOM(scanner, ctx);
-    }
-
-    if (scanner->cancelled) {
-
-      if (scanner->global) {
-        RedisModule_Log(ctx, "notice", "Scanning indexes in background: cancelled (scanned=%zu)",
-                        scanner->scannedKeys);
-      } else {
-        RedisModule_Log(ctx, "notice", "Scanning index %s in background: cancelled (scanned=%zu)",
-                    scanner->spec_name_for_logs, scanner->scannedKeys);
-        goto end;
-      }
-    }
-  }
-
-  if (scanner->isDebug) {
-    DebugIndexesScanner* dScanner = (DebugIndexesScanner*)scanner;
-    dScanner->status = DEBUG_INDEX_SCANNER_CODE_DONE;
-  }
-
-  if (scanner->global) {
-    RedisModule_Log(ctx, "notice", "Scanning indexes in background: done (scanned=%zu)",
-                    scanner->scannedKeys);
-  } else {
-    RedisModule_Log(ctx, "notice", "Scanning index %s in background: done (scanned=%zu)",
-                    scanner->spec_name_for_logs, scanner->scannedKeys);
-  }
-
-end:
-  if (!scanner->cancelled && scanner->global) {
-    Indexes_SetTempSpecsTimers(TimerOp_Add);
-  }
-
-  IndexesScanner_Free(scanner);
-
-  RedisModule_ThreadSafeContextUnlock(ctx);
-  RedisModule_ScanCursorDestroy(cursor);
-  RedisModule_FreeThreadSafeContext(ctx);
-}
-
-//---------------------------------------------------------------------------------------------
-
-static void IndexSpec_ScanAndReindexAsync(StrongRef spec_ref) {
-  if (!reindexPool) {
-    reindexPool = redisearch_thpool_create(1, DEFAULT_HIGH_PRIORITY_BIAS_THRESHOLD, LogCallback, "reindex");
-  }
-#ifdef _DEBUG
-  IndexSpec* spec = (IndexSpec*)StrongRef_Get(spec_ref);
-  const char* indexName = IndexSpec_FormatName(spec, RSGlobalConfig.hideUserDataFromLog);
-  RedisModule_Log(RSDummyContext, "notice", "Register index %s for async scan", indexName);
-#endif
-  IndexesScanner *scanner;
-  if (globalDebugCtx.debugMode) {
-    // If we are in debug mode, we need to allocate a debug scanner
-    scanner = (IndexesScanner*)DebugIndexesScanner_New(spec_ref);
-    // If we need to pause before the scan, we set the pause flag
-    if (globalDebugCtx.bgIndexing.pauseBeforeScan) {
-      globalDebugCtx.bgIndexing.pause = true;
-    }
-  } else {
-    scanner = IndexesScanner_New(spec_ref);
-  }
-
-  redisearch_thpool_add_work(reindexPool, (redisearch_thpool_proc)Indexes_ScanAndReindexTask, scanner, THPOOL_PRIORITY_HIGH);
-}
-
-void ReindexPool_ThreadPoolDestroy() {
-  if (reindexPool != NULL) {
-    RedisModule_ThreadSafeContextUnlock(RSDummyContext);
-    redisearch_thpool_destroy(reindexPool);
-    reindexPool = NULL;
-    RedisModule_ThreadSafeContextLock(RSDummyContext);
-  }
-}
 
 //---------------------------------------------------------------------------------------------
 
@@ -3290,90 +2870,7 @@ void IndexSpec_AddToInfo(RedisModuleInfoCtx *ctx, IndexSpec *sp, bool obfuscate,
   }
 }
 
-// Assumes that the spec is in a safe state to set a scanner on it (write lock or main thread)
-void IndexSpec_ScanAndReindex(RedisModuleCtx *ctx, StrongRef spec_ref) {
-  size_t nkeys = RedisModule_DbSize(ctx);
-  if (nkeys > 0) {
-    IndexSpec_ScanAndReindexAsync(spec_ref);
-  }
-}
 
-// only used on "RDB load finished" event (before the server is ready to accept commands)
-// so it threadsafe
-void IndexSpec_DropLegacyIndexFromKeySpace(IndexSpec *sp) {
-  RedisSearchCtx ctx = SEARCH_CTX_STATIC(RSDummyContext, sp);
-
-  rune *rstr = NULL;
-  t_len slen = 0;
-  float score = 0;
-  int dist = 0;
-  size_t termLen;
-
-  TrieIterator *it = Trie_Iterate(ctx.spec->terms, "", 0, 0, 1);
-  while (TrieIterator_Next(it, &rstr, &slen, NULL, &score, NULL, &dist)) {
-    char *res = runesToStr(rstr, slen, &termLen);
-    RedisModuleString *keyName = Legacy_fmtRedisTermKey(&ctx, res, strlen(res));
-    Redis_LegacyDropScanHandler(ctx.redisCtx, keyName, &ctx);
-    RedisModule_FreeString(ctx.redisCtx, keyName);
-    rm_free(res);
-  }
-  TrieIterator_Free(it);
-
-  // Delete the numeric, tag, and geo indexes which reside on separate keys
-  for (size_t i = 0; i < ctx.spec->numFields; i++) {
-    const FieldSpec *fs = ctx.spec->fields + i;
-    if (FIELD_IS(fs, INDEXFLD_T_NUMERIC)) {
-      RedisModuleString *key = IndexSpec_LegacyGetFormattedKey(ctx.spec, fs, INDEXFLD_T_NUMERIC);
-      Redis_LegacyDeleteKey(ctx.redisCtx, key);
-    }
-    if (FIELD_IS(fs, INDEXFLD_T_TAG)) {
-      RedisModuleString *key = IndexSpec_LegacyGetFormattedKey(ctx.spec, fs, INDEXFLD_T_TAG);
-      Redis_LegacyDeleteKey(ctx.redisCtx, key);
-    }
-    if (FIELD_IS(fs, INDEXFLD_T_GEO)) {
-      RedisModuleString *key = IndexSpec_LegacyGetFormattedKey(ctx.spec, fs, INDEXFLD_T_GEO);
-      Redis_LegacyDeleteKey(ctx.redisCtx, key);
-    }
-  }
-  HiddenString_LegacyDropFromKeySpace(ctx.redisCtx, INDEX_SPEC_KEY_FMT, sp->specName);
-}
-
-void Indexes_UpgradeLegacyIndexes() {
-  dictIterator *iter = dictGetIterator(legacySpecDict);
-  dictEntry *entry = NULL;
-  while ((entry = dictNext(iter))) {
-    StrongRef spec_ref = dictGetRef(entry);
-    IndexSpec *sp = StrongRef_Get(spec_ref);
-    IndexSpec_DropLegacyIndexFromKeySpace(sp);
-
-    // recreate the doctable
-    DocTable_Free(&sp->docs);
-    sp->docs = DocTable_New(INITIAL_DOC_TABLE_SIZE);
-
-    // clear index stats
-    memset(&sp->stats, 0, sizeof(sp->stats));
-    // Init the index error
-    sp->stats.indexError = IndexError_Init();
-
-    // put the new index in the global spec dictionaries (by name and by specId)
-    dictAdd(specDict_g, (void*)sp->specName, spec_ref.rm);
-    dictAdd(specIdDict_g, (void*)(uintptr_t)sp->specId, spec_ref.rm);
-  }
-  dictReleaseIterator(iter);
-}
-
-void Indexes_ScanAndReindex() {
-  if (!reindexPool) {
-    reindexPool = redisearch_thpool_create(1, DEFAULT_HIGH_PRIORITY_BIAS_THRESHOLD, LogCallback, "reindex");
-  }
-
-  RedisModule_Log(RSDummyContext, "notice", "Scanning all indexes");
-  IndexesScanner *scanner = IndexesScanner_NewGlobal();
-  // check no global scan is in progress
-  if (scanner) {
-    redisearch_thpool_add_work(reindexPool, (redisearch_thpool_proc)Indexes_ScanAndReindexTask, scanner, THPOOL_PRIORITY_HIGH);
-  }
-}
 
 ///////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -3438,7 +2935,7 @@ static void IndexSpec_EnsureTagDiskIndexes(IndexSpec *sp) {
 // On the failure path, sp->diskSpec (if already created) is closed by the
 // spec destructor; any unapplied per-field blobs and partially-created
 // vecSimIndexes are released by FieldSpec_Cleanup when the spec is freed.
-static bool IndexSpec_SSTRdbOpenAndApply(RedisModuleCtx *ctx, IndexSpec *sp) {
+bool IndexSpec_SSTRdbOpenAndApply(RedisModuleCtx *ctx, IndexSpec *sp) {
   // The disk layer consumes pendingDiskRdbState by ownership: max_doc_id and
   // deleted_ids are moved into the spec. Per-field vector in-memory state was
   // already deserialized into an unbound VecSimIndex by FieldSpec_RdbLoad and
@@ -3768,7 +3265,7 @@ static int IndexSpec_StoreAfterRdbLoad(IndexSpec *sp) {
   return REDISMODULE_OK;
 }
 
-static int IndexSpec_CreateFromRdb(RedisModuleIO *rdb, int encver, bool useSst, QueryError *status) {
+int IndexSpec_CreateFromRdb(RedisModuleIO *rdb, int encver, bool useSst, QueryError *status) {
   // Load the index spec using the new function
   IndexSpec *sp = IndexSpec_RdbLoad(rdb, encver, useSst, status);
   return IndexSpec_StoreAfterRdbLoad(sp);
@@ -3919,68 +3416,8 @@ void IndexSpec_LegacyRdbSave(RedisModuleIO *rdb, void *value) {
   return;
 }
 
-int Indexes_RdbLoad(RedisModuleIO *rdb, int encver, int when) {
-  const bool useSst = CheckRdbSstPersistence(RedisModule_GetContextFromIO(rdb), "RDB Load");
-  size_t nIndexes = 0;
-  QueryError status = QueryError_Default();
 
-  if (encver < INDEX_MIN_COMPAT_VERSION) {
-    return REDISMODULE_ERR;
-  }
 
-  nIndexes = LoadUnsigned_IOError(rdb, goto cleanup);
-
-  if (unlikely(nIndexes > maxIndexes_g)) {
-    RedisModule_LogIOError(
-        rdb, "warning",
-        "RDB Load: Number of indexes (%zu) exceeds maximum allowed (%u)",
-        nIndexes, maxIndexes_g);
-    return REDISMODULE_ERR;
-  }
-  if (!SearchDisk_CheckLimitNumberOfIndexes(nIndexes)) {
-    RedisModule_LogIOError(rdb, "warning", "Too many indexes for flex. Having %zu indexes, but flex only supports %d.", nIndexes, FLEX_MAX_INDEX_COUNT);
-    return REDISMODULE_ERR;
-  }
-  for (size_t i = 0; i < nIndexes; ++i) {
-    if (IndexSpec_CreateFromRdb(rdb, encver, useSst, &status) != REDISMODULE_OK) {
-      RedisModule_LogIOError(rdb, "warning", "RDB Load: %s", QueryError_GetDisplayableError(&status, RSGlobalConfig.hideUserDataFromLog));
-      QueryError_ClearError(&status);
-      return REDISMODULE_ERR;
-    }
-  }
-
-  // If we have indexes in the auxiliary data, we need to subscribe to the
-  // keyspace notifications
-  Initialize_KeyspaceNotifications();
-
-  return REDISMODULE_OK;
-
-cleanup:
-  return REDISMODULE_ERR;
-}
-
-void Indexes_RdbSave(RedisModuleIO *rdb, int when) {
-  RedisModuleCtx *ctx = RedisModule_GetContextFromIO(rdb);
-  const int contextFlags = RedisModule_GetContextFlags(ctx);
-
-  RedisModule_SaveUnsigned(rdb, dictSize(specDict_g));
-
-  dictIterator *iter = dictGetIterator(specDict_g);
-  dictEntry *entry = NULL;
-  while ((entry = dictNext(iter))) {
-    StrongRef spec_ref = dictGetRef(entry);
-    IndexSpec *sp = StrongRef_Get(spec_ref);
-    IndexSpec_RdbSave(rdb, sp, contextFlags);
-  }
-
-  dictReleaseIterator(iter);
-}
-
-void Indexes_RdbSave2(RedisModuleIO *rdb, int when) {
-  if (dictSize(specDict_g)) {
-    Indexes_RdbSave(rdb, when);
-  }
-}
 
 
 void *IndexSpec_RdbLoad_Logic(RedisModuleIO *rdb, int encver) {
@@ -4052,23 +3489,6 @@ int CompareVersions(Version v1, Version v2) {
   return 0;
 }
 
-void Indexes_Propagate(RedisModuleCtx *ctx) {
-  dictIterator *iter = dictGetIterator(specDict_g);
-  dictEntry *entry;
-  while ((entry = dictNext(iter))) {
-    StrongRef spec_ref = dictGetRef(entry);
-    IndexSpec *sp = StrongRef_Get(spec_ref);
-    RS_ASSERT(sp != NULL);
-    RedisModuleString *serialized = IndexSpec_Serialize(sp);
-    RS_ASSERT(serialized != NULL);
-    int rc = RedisModule_ClusterPropagateForSlotMigration(ctx, CMD_FOR_ENV(RS_RESTORE_IF_NX), "cls", SPEC_SCHEMA_STR, INDEX_CURRENT_VERSION, serialized);
-    if (rc != REDISMODULE_OK) {
-      RedisModule_Log(ctx, "warning", "Failed to propagate index '%s' during slot migration. errno: %d", IndexSpec_FormatName(sp, RSGlobalConfig.hideUserDataFromLog), errno);
-    }
-    RedisModule_FreeString(NULL, serialized);
-  }
-  dictReleaseIterator(iter);
-}
 
 static void IndexSpec_RdbSave_Wrapper(RedisModuleIO *rdb, void *value) {
   RedisModuleCtx *ctx = RedisModule_GetContextFromIO(rdb);
@@ -4272,551 +3692,20 @@ void IndexSpec_DeleteDocById(IndexSpec *spec, t_docId docId) {
 
 ///////////////////////////////////////////////////////////////////////////////////////////////
 
-static void onFlush(RedisModuleCtx *ctx, RedisModuleEvent eid, uint64_t subevent, void *data) {
-  if (subevent != REDISMODULE_SUBEVENT_FLUSHDB_START) {
-    return;
-  }
-  if (specDict_g) {
-    Indexes_Free(ctx, specDict_g, true);
-    // specDict_g itself is not actually freed
-  }
-  Dictionary_Clear();
-  RSGlobalStats.totalStats.used_dialects = 0;
-}
 
-void Indexes_Init(RedisModuleCtx *ctx) {
-  if (!specDict_g) {
-    specDict_g = dictCreate(&dictTypeHeapHiddenStrings, NULL);
-  }
-  if (!specIdDict_g) {
-    specIdDict_g = dictCreate(&dictTypeUint64, NULL);
-  }
-  RedisModule_SubscribeToServerEvent(ctx, RedisModuleEvent_FlushDB, onFlush);
-  SchemaPrefixes_Create();
-}
 
-size_t Indexes_Count() {
-  return dictSize(specDict_g);
-}
 
-SpecOpIndexingCtx *Indexes_FindMatchingSchemaRules(RedisModuleCtx *ctx, RedisModuleString *key,
-                                                   DocumentType type, bool runFilters,
-                                                   RedisModuleString *keyToReadData) {
-  if (!keyToReadData) {
-    keyToReadData = key;
-  }
-  SpecOpIndexingCtx *res = rm_malloc(sizeof(*res));
-  res->specs = dictCreate(&dictTypeHeapHiddenStrings, NULL);
-  res->specsOps = array_new(SpecOpCtx, 10);
-  if (dictSize(specDict_g) == 0) {
-    return res;
-  }
-  dict *specs = res->specs;
 
-#if defined(_DEBUG) && 0
-  RLookupKey *k = RLookup_GetKey_LoadEx(&r->lk, UNDERSCORE_KEY, strlen(UNDERSCORE_KEY), UNDERSCORE_KEY, RLOOKUP_F_NOFLAGS);
-  RSValue *v = RLookup_GetItem(k, &r->row);
-  const char *x = RSValue_StringPtrLen(v, NULL);
-  RedisModule_Log(RSDummyContext, "notice", "Indexes_FindMatchingSchemaRules: x=%s", x);
-  const char *f = "name";
-  k = RLookup_GetKey_ReadEx(&r->lk, f, strlen(f), RLOOKUP_F_NOFLAGS);
-  if (k) {
-    v = RLookup_GetItem(k, &r->row);
-    x = RSValue_StringPtrLen(v, NULL);
-  }
-#endif  // _DEBUG
 
-  size_t n;
-  const char *key_p = RedisModule_StringPtrLen(key, &n);
-  // collect specs that their name is prefixed by the key name
-  // `prefixes` includes list of arrays of specs, one for each prefix of key name
-  TrieMapResultBuf prefixes = TrieMap_FindPrefixes(SchemaPrefixes_g, key_p, n);
-  for (int i = 0; i < TrieMapResultBuf_Len(&prefixes); ++i) {
-    SchemaPrefixNode *node = TrieMapResultBuf_GetByIndex(&prefixes, i);
-    for (int j = 0; j < array_len(node->index_specs); ++j) {
-      StrongRef global = node->index_specs[j];
-      IndexSpec *spec = StrongRef_Get(global);
-      if (spec && !dictFind(specs, spec->specName)) {
-        // skip if document type does not match the index type
-        // The unsupported type is needed for crdt empty keys (deleted)
-        if (type != DocumentType_Unsupported && type != spec->rule->type) {
-          continue;
-        }
 
-        SpecOpCtx specOp = {
-            .spec = spec,
-            .op = SpecOp_Add,
-        };
-        array_append(res->specsOps, specOp);
-        dictEntry *entry = dictAddRaw(specs, (void*)spec->specName, NULL);
-        // put the location on the specsOps array so we can get it
-        // fast using index name
-        entry->v.u64 = array_len(res->specsOps) - 1;
-      }
-    }
-  }
-  TrieMapResultBuf_Free(prefixes);
 
-  if (runFilters) {
-    // We load the data from the `keyToReadData` key, which is the key the old
-    // key was changed to, since the old key is already deleted.
-    key_p = RedisModule_StringPtrLen(keyToReadData, NULL);
 
-    EvalCtx *r = NULL;
-    for (size_t i = 0; i < array_len(res->specsOps); ++i) {
-      SpecOpCtx *specOp = res->specsOps + i;
-      IndexSpec *spec = specOp->spec;
-      if (!spec->rule->filter_exp) {
-        continue;
-      }
 
-      // load document only if required
-      if (!r) {
-        r = EvalCtx_Create(EVAL_MODE_INDEX);
-      }
 
-      RedisSearchCtx sctx = { .redisCtx = ctx };
-      QueryError status = QueryError_Default();
-      RLookup_LoadRuleFields(&sctx, &r->lk, &r->row, spec, key_p, &status);
-      QueryError_ClearError(&status); // TODO: report errors
 
-      if (!SchemaRule_FilterPasses(r, spec->rule->filter_exp)) {
-        if (dictFind(specs, spec->specName)) {
-          specOp->op = SpecOp_Del;
-        }
-      }
-      // Clean up state between iterations (indexes)
-      QueryError_ClearError(&r->status);
-      RLookup_Cleanup(&r->lk);
-      r->lk = RLookup_New();
-      RLookupRow_Reset(&r->row);
-    }
 
-    if (r) {
-      EvalCtx_Destroy(r);
-    }
-  }
-  return res;
-}
-
-static bool hashFieldChanged(IndexSpec *spec, RedisModuleString **hashFields) {
-  if (hashFields == NULL) {
-    return true;
-  }
-
-  // TODO: improve implementation to avoid O(n^2)
-  for (size_t i = 0; hashFields[i] != NULL; ++i) {
-    size_t length = 0;
-    const char *field = RedisModule_StringPtrLen(hashFields[i], &length);
-    for (size_t j = 0; j < spec->numFields; ++j) {
-      if (!HiddenString_CompareC(spec->fields[j].fieldName, field, length)) {
-        return true;
-      }
-    }
-    // optimize. change of score and payload fields just require an update of the doc table
-    if ((spec->rule->lang_field && !strcmp(field, spec->rule->lang_field)) ||
-        (spec->rule->score_field && !strcmp(field, spec->rule->score_field)) ||
-        (spec->rule->payload_field && !strcmp(field, spec->rule->payload_field))) {
-      return true;
-    }
-  }
-  return false;
-}
-
-void Indexes_SpecOpsIndexingCtxFree(SpecOpIndexingCtx *specs) {
-  dictRelease(specs->specs);
-  array_free(specs->specsOps);
-  rm_free(specs);
-}
-
-void Indexes_UpdateMatchingWithSchemaRules(RedisModuleCtx *ctx, RedisModuleString *key, DocumentType type,
-                                           RedisModuleString **hashFields) {
-  if (type == DocumentType_Unsupported) {
-    // COPY could overwrite a hash/json with other types so we must try and remove old doc.
-    Indexes_DeleteMatchingWithSchemaRules(ctx, key, type, hashFields);
-    return;
-  }
-
-  SpecOpIndexingCtx *specs = Indexes_FindMatchingSchemaRules(ctx, key, type, true, NULL);
-
-  for (size_t i = 0; i < array_len(specs->specsOps); ++i) {
-    SpecOpCtx *specOp = specs->specsOps + i;
-
-    if (hashFieldChanged(specOp->spec, hashFields)) {
-      if (specOp->op == SpecOp_Add) {
-        IndexSpec_UpdateDoc(specOp->spec, ctx, key, type);
-      } else {
-        // specOp->op is SpecOp_Del when the key matches the index prefix but
-        // the filter expression fails (e.g. a field value changed so the filter
-        // no longer passes, or a required field is missing). If the document was
-        // previously indexed, it must be removed now.
-        IndexSpec_DeleteDoc(specOp->spec, ctx, key);
-      }
-    }
-  }
-
-  Indexes_SpecOpsIndexingCtxFree(specs);
-}
-
-void Indexes_UpdateMatchingDocExpiration(RedisModuleCtx *ctx, RedisModuleString *key, DocumentType type) {
-  if (type == DocumentType_Unsupported || !RSGlobalConfig.monitorExpiration) {
-    return;
-  }
-
-  // Find all indexes that match the key's name prefix. runFilters=false:
-  // EXPIRE/PERSIST do not change field values, so a doc's schema-rule FILTER
-  // outcome cannot have flipped since the last write. The DocTable lookup
-  // below (DocTable_BorrowByKeyR) is the only discriminator we need — it
-  // returns the existing DMD when the doc is currently indexed, and NULL
-  // otherwise. Re-evaluating the filter here would re-read hash/JSON fields
-  // on the main thread for a result the loop would ignore.
-  SpecOpIndexingCtx *specs = Indexes_FindMatchingSchemaRules(ctx, key, type, false, NULL);
-
-  // EXPIRE/PERSIST notifications fire for every key in the keyspace. When no
-  // spec prefix matches we have nothing to update, so skip the OpenKey/TTL
-  // read entirely — that overhead would otherwise be paid on every event.
-  if (array_len(specs->specsOps) == 0) {
-    Indexes_SpecOpsIndexingCtxFree(specs);
-    return;
-  }
-
-  RedisModuleKey *kp = RedisModule_OpenKey(ctx, key, DOCUMENT_OPEN_KEY_INDEXING_FLAGS);
-  RS_ASSERT(kp);
-  t_expirationTimePoint ttl = GetKeyExpirationTime(kp);
-  RedisModule_CloseKey(kp);
-
-  for (size_t i = 0; i < array_len(specs->specsOps); ++i) {
-    SpecOpCtx *specOp = &specs->specsOps[i];
-    IndexSpec *spec = specOp->spec;
-    // Skip TTL update for specs that opted out of doc-level TTL tracking,
-    // matching the gate in Document_LoadSchemaFieldHash/Json. Otherwise
-    // DocTable_IsDocExpired could drop rows for an index that doesn't
-    // monitor TTLs.
-    //
-    // No spec lock needed: monitorDocumentExpiration is only written from
-    // main-thread callbacks (spec init, FT.CONFIG SET, FT.DEBUG
-    // MONITOR_EXPIRATION), and this keyspace-notification callback also
-    // runs on the main thread, so the Redis event loop serializes them.
-    // The spec write lock guards index data against background workers,
-    // not main-thread config flags (same pattern as the load path).
-    if (!spec->monitorDocumentExpiration) {
-      continue;
-    }
-    RedisSearchCtx sctx = SEARCH_CTX_STATIC(ctx, spec);
-    // Read lock is sufficient: the only mutation is the relaxed atomic store on
-    // `dmd->expirationTimeNs` inside DocTable_SetDocExpiration (paired with a
-    // relaxed atomic load in DocTable_IsDocExpired). The DMD chain traversal
-    // and refcount manipulation in DocTable_BorrowByKeyR / DMD_Return are
-    // explicitly documented as safe under either lock mode (doc_table.c, near
-    // DocTable_GetOwn). Concurrent writers cannot race here because keyspace
-    // notifications all dispatch on the Redis main thread, so this callback is
-    // serialized against itself and against other notification-driven writers
-    // by the event loop, not by the spec lock.
-    RedisSearchCtx_LockSpecRead(&sctx);
-    const RSDocumentMetadata *cdmd = DocTable_BorrowByKeyR(&spec->docs, key);
-    if (cdmd) {
-      // Only the doc-level TTL changes here. EXPIRE/PERSIST do not affect
-      // HEXPIRE state, and the per-field TTL table must be left untouched —
-      // mutating it would also require the spec write lock, which we do not
-      // hold on this fast path.
-      DocTable_SetDocExpiration((RSDocumentMetadata *)cdmd, ttl);
-      DMD_Return(cdmd);
-    }
-    RedisSearchCtx_UnlockSpec(&sctx);
-  }
-
-  Indexes_SpecOpsIndexingCtxFree(specs);
-}
-
-void Indexes_DeleteMatchingWithSchemaRules(RedisModuleCtx *ctx, RedisModuleString *key,
-                                           DocumentType type,
-                                           RedisModuleString **hashFields) {
-  SpecOpIndexingCtx *specs = Indexes_FindMatchingSchemaRules(ctx, key, type, false, NULL);
-
-  for (size_t i = 0; i < array_len(specs->specsOps); ++i) {
-    SpecOpCtx *specOp = specs->specsOps + i;
-    if (hashFieldChanged(specOp->spec, hashFields)) {
-      IndexSpec_DeleteDoc(specOp->spec, ctx, key);
-    }
-  }
-
-  Indexes_SpecOpsIndexingCtxFree(specs);
-}
-
-// True iff the spec has any field with INDEXMISSING. Linear scan over the
-// schema's fields[] array; called from the HEXPIRE fast path on the main
-// thread without the spec lock, which is safe because the schema descriptors
-// read here are only mutated by FT.CREATE / FT.ALTER / RDB load on the same
-// thread.
-static bool specHasIndexMissing(const IndexSpec *spec) {
-  for (size_t i = 0; i < spec->numFields; ++i) {
-    if (FieldSpec_IndexesMissing(&spec->fields[i])) {
-      return true;
-    }
-  }
-  return false;
-}
-
-void Indexes_UpdateMatchingHashFieldExpiration(RedisModuleCtx *ctx, RedisModuleString *key,
-                                               DocumentType type) {
-
-  if (type == DocumentType_Unsupported) {
-    return;
-  }
-  RS_ASSERT(type == DocumentType_Hash);
-
-  SpecOpIndexingCtx *specs = Indexes_FindMatchingSchemaRules(ctx, key, DocumentType_Hash, false, NULL);
-
-  // HEXPIRE/HPERSIST notifications fire for every hash in the keyspace. When
-  // no spec prefix matches we have nothing to update, so skip the OpenKey /
-  // HashFieldMinExpire read entirely — that overhead would otherwise be paid
-  // on every event on non-indexed hashes.
-  if (array_len(specs->specsOps) == 0) {
-    Indexes_SpecOpsIndexingCtxFree(specs);
-    return;
-  }
-
-  RedisModuleKey *k = RedisModule_OpenKey(ctx, key, DOCUMENT_OPEN_KEY_INDEXING_FLAGS);
-  RS_ASSERT(k);
-
-  // Hash-level gate: if no field on this hash has a TTL, every per-spec
-  // FieldExpiration array is empty, so we can skip the per-field HashGet
-  // pass entirely. The Redis hash is owned by the main thread, so this read
-  // is safe without the spec lock.
-  const bool hashHasAnyFieldExpire =
-      (RedisModule_HashFieldMinExpire(k) != REDISMODULE_NO_EXPIRE);
-
-  for (size_t i = 0; i < array_len(specs->specsOps); ++i) {
-    SpecOpCtx *specOp = &specs->specsOps[i];
-    IndexSpec *spec = specOp->spec;
-    // Specs that opted out of HFE tracking have no TTL table state to refresh
-    // and HEXPIRE cannot otherwise affect their indexed view of the doc.
-    //
-    // No spec lock needed for the read: monitorFieldExpiration is only
-    // written from main-thread callbacks, and this keyspace-notification
-    // callback also runs on the main thread, so the Redis event loop
-    // serializes them. The spec write lock guards index data against
-    // background workers, not main-thread config flags.
-    if (!spec->monitorFieldExpiration) {
-      continue;
-    }
-
-    // INDEXMISSING relies on reindexing: the missing-docs iterator walks
-    // an inverted index that requires monotonically increasing docIds, so
-    // we cannot patch it from the fast path. Fall back to full reindex.
-    //
-    // Indexes_FindMatchingSchemaRules was called with runFilters=false, so
-    // the schema FILTER was not evaluated. Re-evaluate it here via
-    // SchemaRule_ShouldIndex so an HEXPIRE on a hash whose PREFIX matches
-    // but whose FILTER rejects it does not incorrectly add the doc to the
-    // index. Matches the SpecOp_Add / SpecOp_Del split the slow path
-    // produces in Indexes_UpdateMatchingWithSchemaRules.
-    if (specHasIndexMissing(spec)) {
-      if (SchemaRule_ShouldIndex(spec, key, type)) {
-        IndexSpec_UpdateDoc(spec, ctx, key, type);
-      } else {
-        IndexSpec_DeleteDoc(spec, ctx, key);
-      }
-      continue;
-    }
-
-    // Build the FieldExpiration array lock-free. Two reads happen outside
-    // the spec write lock here:
-    //   1. spec->fields / spec->numFields — only mutated by FT.CREATE,
-    //      FT.ALTER, and RDB load, all of which run on the main thread.
-    //      This keyspace-notification callback also runs on the main
-    //      thread, so the Redis event loop serializes them; the schema
-    //      cannot change between iterations here. (Same invariant as
-    //      specHasIndexMissing above.)
-    //   2. The Redis hash — owned by the main thread and not guarded by
-    //      the spec lock at all; the spec rwlock protects index state,
-    //      not the Redis keyspace.
-    arrayof(FieldExpiration) sorted = NULL;
-    if (hashHasAnyFieldExpire) {
-      for (size_t ii = 0; ii < spec->numFields; ++ii) {
-        Document_LoadHashFieldExpiration(k, &spec->fields[ii], ii, &sorted);
-      }
-    }
-
-    RedisSearchCtx sctx = SEARCH_CTX_STATIC(ctx, spec);
-    RedisSearchCtx_LockSpecWrite(&sctx);
-
-    const RSDocumentMetadata *cdmd = DocTable_BorrowByKeyR(&spec->docs, key);
-    if (cdmd) {
-      // Hands ownership of `sorted` to the doc table (or frees it if empty).
-      DocTable_UpdateFieldExpiration(&spec->docs, (RSDocumentMetadata *)cdmd, sorted);
-      sorted = NULL;
-      DMD_Return(cdmd);
-    }
-
-    RedisSearchCtx_UnlockSpec(&sctx);
-
-    // Doc not in this index (filter failed or never indexed): free the array
-    // we built speculatively. `array_free` is NULL-safe.
-    array_free(sorted);
-  }
-
-  RedisModule_CloseKey(k);
-  Indexes_SpecOpsIndexingCtxFree(specs);
-}
-
-void Indexes_ReplaceMatchingWithSchemaRules(RedisModuleCtx *ctx, RedisModuleString *from_key,
-                                            RedisModuleString *to_key) {
-  DocumentType type = getDocTypeFromString(to_key);
-  if (type == DocumentType_Unsupported) {
-    return;
-  }
-
-  SpecOpIndexingCtx *from_specs = Indexes_FindMatchingSchemaRules(ctx, from_key, type, true, to_key);
-  SpecOpIndexingCtx *to_specs = Indexes_FindMatchingSchemaRules(ctx, to_key, type, true, NULL);
-
-  size_t from_len, to_len;
-  const char *from_str = RedisModule_StringPtrLen(from_key, &from_len);
-  const char *to_str = RedisModule_StringPtrLen(to_key, &to_len);
-
-  // Handle specs that match the old key (whether they match the new key or not)
-  for (size_t i = 0; i < array_len(from_specs->specsOps); ++i) {
-    SpecOpCtx *specOp = from_specs->specsOps + i;
-    IndexSpec *spec = specOp->spec;
-    if (specOp->op == SpecOp_Del) {
-      // the document is not in the index from the first place
-      continue;
-    }
-    dictEntry *entry = dictFind(to_specs->specs, spec->specName);
-    if (entry) {
-      // The document should be indexed by the new key as well, so we need to update the key name in the index.
-      RedisSearchCtx sctx = SEARCH_CTX_STATIC(ctx, spec);
-      RedisSearchCtx_LockSpecWrite(&sctx);
-
-      // Perform the rename
-      if (SearchDisk_IsEnabled()) {
-        uint64_t docId;
-        // After RENAME, the metadata lives on to_key (rename callback keeps it).
-        if (DocIdMeta_Get(ctx, to_key, spec->specId, &docId) == REDISMODULE_OK) {
-          // Update the key name in the disk doc table
-          SearchDisk_ReplaceKey(spec->diskSpec, docId, to_str, to_len);
-        }
-      } else {
-        DocTable_Replace(&spec->docs, from_str, from_len, to_str, to_len);
-      }
-
-      RedisSearchCtx_UnlockSpec(&sctx);
-      size_t index = entry->v.u64;
-      dictDelete(to_specs->specs, spec->specName);
-      array_del_fast(to_specs->specsOps, index);
-    } else {
-      // The document should not be indexed by the new key, so we need to delete the old document from the index.
-      if (SearchDisk_IsEnabled()) {
-        // After RENAME, from_key no longer exists. The metadata is on to_key.
-        // Look up the docId from to_key's metadata and delete by id.
-        uint64_t docId;
-        if (DocIdMeta_Get(ctx, to_key, spec->specId, &docId) == REDISMODULE_OK) {
-          IndexSpec_DeleteDocById(spec, (t_docId)docId);
-          DocIdMeta_Delete(ctx, to_key, spec->specId);
-        }
-      } else {
-        // For RAM case, look up by old key name and delete
-        IndexSpec_DeleteDoc(spec, ctx, from_key);
-      }
-    }
-  }
-
-  // Handle specs that didn't match the old key but match the new key
-  for (size_t i = 0; i < array_len(to_specs->specsOps); ++i) {
-    SpecOpCtx *specOp = to_specs->specsOps + i;
-    if (specOp->op == SpecOp_Del) {
-      // not need to index
-      // also no need to delete because we know that the document is
-      // not in the index because if it was there we would handle it
-      // on the spec from section.
-      continue;
-    }
-    IndexSpec_UpdateDoc(specOp->spec, ctx, to_key, type);
-  }
-  Indexes_SpecOpsIndexingCtxFree(from_specs);
-  Indexes_SpecOpsIndexingCtxFree(to_specs);
-}
-
-void Indexes_List(RedisModule_Reply* reply, bool obfuscate) {
-  RedisModule_Reply_Set(reply);
-  dictIterator *iter = dictGetIterator(specDict_g);
-  dictEntry *entry = NULL;
-  while ((entry = dictNext(iter))) {
-    StrongRef ref = dictGetRef(entry);
-    IndexSpec *sp = StrongRef_Get(ref);
-    CurrentThread_SetIndexSpec(ref);
-    const char *specName = IndexSpec_FormatName(sp, obfuscate);
-    REPLY_SIMPLE_SAFE(specName);
-    CurrentThread_ClearIndexSpec();
-  }
-  dictReleaseIterator(iter);
-  RedisModule_Reply_SetEnd(reply);
-  RedisModule_EndReply(reply);
-}
 ///////////////////////////////////////////////////////////////////////////////////////////////
 
-// Debug Scanner Functions
-
-static DebugIndexesScanner *DebugIndexesScanner_New(StrongRef global_ref) {
-
-  DebugIndexesScanner *dScanner = rm_realloc(IndexesScanner_New(global_ref), sizeof(DebugIndexesScanner));
-
-  dScanner->maxDocsTBscanned = globalDebugCtx.bgIndexing.maxDocsTBscanned;
-  dScanner->maxDocsTBscannedPause = globalDebugCtx.bgIndexing.maxDocsTBscannedPause;
-  dScanner->wasPaused = false;
-  dScanner->status = DEBUG_INDEX_SCANNER_CODE_NEW;
-  dScanner->base.isDebug = true;
-  dScanner->pauseOnOOM = globalDebugCtx.bgIndexing.pauseOnOOM;
-  dScanner->pauseBeforeOOMRetry = globalDebugCtx.bgIndexing.pauseBeforeOOMretry;
-
-  IndexSpec *spec = StrongRef_Get(global_ref);
-  spec->scanner = (IndexesScanner*)dScanner;
-
-  return dScanner;
-}
-
-static void DebugIndexes_ScanProc(RedisModuleCtx *ctx, RedisModuleString *keyname, RedisModuleKey *key,
-                             DebugIndexesScanner *dScanner) {
-
-  IndexesScanner *scanner = &(dScanner->base);
-
-  if (dScanner->status == DEBUG_INDEX_SCANNER_CODE_NEW) {
-    dScanner->status = DEBUG_INDEX_SCANNER_CODE_RUNNING;
-  }
-
-  if (dScanner->maxDocsTBscannedPause > 0 && (!dScanner->wasPaused) && scanner->scannedKeys == dScanner->maxDocsTBscannedPause) {
-    globalDebugCtx.bgIndexing.pause = true;
-    dScanner->wasPaused = true;
-  }
-
-  if ((dScanner->maxDocsTBscanned > 0) && (scanner->scannedKeys == dScanner->maxDocsTBscanned)) {
-    scanner->cancelled = true;
-    dScanner->status = DEBUG_INDEX_SCANNER_CODE_CANCELLED;
-  }
-
-  // Check if we need to pause the scan before we release the GIL
-  if (globalDebugCtx.bgIndexing.pause)
-  {
-      // Warning: This section is highly unsafe. RM_Scan does not permit the callback
-      // function (i.e., this function) to release the GIL.
-      // If the key currently being scanned is deleted after the GIL is released,
-      // it can lead to a use-after-free and crash Redis.
-      RedisModule_Log(ctx, "warning", "RM_Scan callback function is releasing the GIL, which is unsafe.");
-
-      RedisModule_ThreadSafeContextUnlock(ctx);
-      while (globalDebugCtx.bgIndexing.pause) { // volatile variable
-        dScanner->status = DEBUG_INDEX_SCANNER_CODE_PAUSED;
-        usleep(1000);
-      }
-      RedisModule_ThreadSafeContextLock(ctx);
-  }
-
-  if (dScanner->status == DEBUG_INDEX_SCANNER_CODE_PAUSED) {
-    dScanner->status = DEBUG_INDEX_SCANNER_CODE_RESUMED;
-  }
-
-  Indexes_ScanProc(ctx, keyname, key, &(dScanner->base));
-}
 
 StrongRef IndexSpecRef_Promote(WeakRef ref) {
   StrongRef strong = WeakRef_Promote(ref);
@@ -4832,116 +3721,11 @@ void IndexSpecRef_Release(StrongRef ref) {
   StrongRef_Release(ref);
 }
 
-// If this function is called, it means that the scan did not complete due to OOM, should be verified by the caller
-static inline void DebugIndexesScanner_pauseCheck(DebugIndexesScanner* dScanner, RedisModuleCtx *ctx, bool pauseField, DebugIndexScannerCode code) {
-  if (!dScanner || !pauseField) {
-    return;
-  }
-  globalDebugCtx.bgIndexing.pause = true;
-  RedisModule_ThreadSafeContextUnlock(ctx);
-  while (globalDebugCtx.bgIndexing.pause) { // volatile variable
-    dScanner->status = code;
-    usleep(1000);
-  }
-  dScanner->status = DEBUG_INDEX_SCANNER_CODE_RESUMED;
-  RedisModule_ThreadSafeContextLock(ctx);
-}
 
-void Indexes_StartRDBLoadingEvent(RedisModuleCtx* ctx) {
-  Indexes_Free(ctx, specDict_g, false);
-  if (!SearchDisk_IsEnabled()) {
-    if (legacySpecDict) {
-      dictEmpty(legacySpecDict, NULL);
-    } else {
-      legacySpecDict = dictCreate(&dictTypeHeapHiddenStrings, NULL);
-    }
-  }
-  g_isLoading = true;
-}
 
-void Indexes_EndRDBLoadingEvent(RedisModuleCtx *ctx) {
-  int hasLegacyIndexes = dictSize(legacySpecDict);
-  Indexes_UpgradeLegacyIndexes();
 
-  // we do not need the legacy dict specs anymore
-  dictRelease(legacySpecDict);
-  legacySpecDict = NULL;
 
-  LegacySchemaRulesArgs_Free(ctx);
 
-  if (hasLegacyIndexes) {
-    Indexes_ScanAndReindex();
-  }
-}
-
-void Indexes_EndLoading() {
-  g_isLoading = false;
-}
-
-// Replica-side SST replication: Open with the pending RDB State every spec
-// that was staged in this round.
-void Indexes_FinishSSTReplication(RedisModuleCtx *ctx) {
-  RS_ASSERT(SearchDisk_IsEnabled());
-  dictIterator *iter = dictGetIterator(specDict_g);
-  dictEntry *entry = NULL;
-  while ((entry = dictNext(iter))) {
-    StrongRef spec_ref = dictGetRef(entry);
-    IndexSpec *sp = StrongRef_Get(spec_ref);
-    RS_ASSERT(sp);
-    RS_ASSERT(sp->pendingDiskRdbState);
-    RS_ASSERT(sp->diskSpec == NULL);
-    RS_ASSERT(!sp->diskRegistered);
-    bool ok = IndexSpec_SSTRdbOpenAndApply(ctx, sp);
-    RS_LOG_ASSERT_ALWAYS(ok, "SST replication: failed to open disk index for spec during LOADING_ENDED");
-    // GC start was deferred by IndexSpec_StoreAfterRdbLoad for the SST path
-    // (diskSpec was NULL there); start it now that the disk handle exists.
-    IndexSpec_StartGC(spec_ref, sp, GCPolicy_Disk);
-  }
-  dictReleaseIterator(iter);
-}
-
-// Replica-side SST replication: abort the in-progress replication round.
-//
-// Drops every spec that was built or staged in this round, freeing any
-// pending disk RDB state and closing or unregistering+closing the disk index
-// as appropriate. Specs at every replication phase are torn down — the
-// replica is left as if replication had never started, and Flex will retry
-// from scratch on the next attempt.
-void Indexes_AbortSSTReplicationLoading(RedisModuleCtx *ctx) {
-  RS_ASSERT(SearchDisk_IsEnabled());
-  if (dictSize(specDict_g) == 0) return;
-
-  RedisModule_Log(ctx, "warning",
-                  "SST replication aborted; tearing down %lu staged index(es)",
-                  (unsigned long)dictSize(specDict_g));
-
-  // Snapshot the refs first since IndexSpec_RemoveFromGlobals mutates specDict_g.
-  arrayof(StrongRef) specs = array_new(StrongRef, dictSize(specDict_g));
-  dictIterator *iter = dictGetIterator(specDict_g);
-  dictEntry *entry = NULL;
-  while ((entry = dictNext(iter))) {
-    array_append(specs, (StrongRef)dictGetRef(entry));
-  }
-  dictReleaseIterator(iter);
-
-  for (size_t i = 0; i < array_len(specs); ++i) {
-    IndexSpec *spec = StrongRef_Get(specs[i]);
-    if (!spec) continue;
-    // Unregister here (not in the destructor) because the destructor may run
-    // on a background thread when the last StrongRef drops, and unregister
-    // needs the main-thread RedisModuleCtx. Must precede the close that
-    // IndexSpec_FreeUnlinkedData performs. SearchDisk_UnregisterIndex is
-    // idempotent, so it safely handles specs aborted before LOADING_ENDED
-    // registered them.
-    if (spec->diskSpec) {
-      SearchDisk_UnregisterIndex(ctx, spec);
-    }
-    // pendingDiskRdbState and diskSpec are freed by IndexSpec_FreeUnlinkedData
-    // once the last StrongRef is dropped below.
-    IndexSpec_RemoveFromGlobals(specs[i], false);
-  }
-  array_free(specs);
-}
 
 // =============================================================================
 // Compaction FFI Functions (called by Rust during GC)
