+++ tests/cpptests/common.h
@@ -12,6 +12,7 @@
 #include "redismock/redismock.h"
 #include "redismock/util.h"
 #include "spec.h"
+#include "indexes.h"
 #include "document.h"
 #include "llapi_test_helpers.h"
 
@@ -61,7 +62,7 @@ template <typename... Ts>
 IndexSpec *createIndex(RedisModuleCtx *ctx, const char *name, Ts... args) {
   RMCK::ArgvList argv("FT.CREATE", name, args...);
   QueryError err{QueryErrorCode(0)};
-  IndexSpec *sp = IndexSpec_CreateNew(ctx, argv, argv.size(), &err);
+  IndexSpec *sp = Indexes_CreateNew(ctx, argv, argv.size(), &err);
   if (!sp) {
     abort();
   }
