+++ tests/cpptests/test_cpp_agg.cpp
@@ -38,7 +38,7 @@ TEST_F(AggTest, testBasic) {
   RMCK::ArgvList args(ctx, "FT.CREATE", "idx", "ON", "HASH",
                       "SCHEMA", "t1", "TEXT", "SORTABLE", "t2", "NUMERIC",
                       "sortable", "t3", "TEXT");
-  auto spec = IndexSpec_CreateNew(ctx, args, args.size(), &qerr);
+  auto spec = Indexes_CreateNew(ctx, args, args.size(), &qerr);
   ASSERT_TRUE(spec);
 
   // Try to create a document...
