+++ tests/cpptests/test_cpp_document.cpp
@@ -11,6 +11,7 @@
 #include "redismock/redismock.h"
 #include "redismock/util.h"
 #include "document.h"
+#include "indexes.h"
 #include "llapi_test_helpers.h"
 
 class DocumentTest : public ::testing::Test {
@@ -82,7 +83,7 @@ TEST_F(DocumentTest, testLoadSchema) {
   // Create a database
   QueryError status = QueryError_Default();
   RMCK::ArgvList args(ctx, "FT.CREATE", "idx", "ON", "HASH", "SCHEMA", "t1", "TEXT", "t2", "TEXT");
-  auto spec = IndexSpec_CreateNew(ctx, args, args.size(), &status);
+  auto spec = Indexes_CreateNew(ctx, args, args.size(), &status);
   ASSERT_FALSE(spec == NULL);
 
   Document d = {0};
